Add files via upload
In this part, i evaluate the model perforance using Gini and kolmogorov-smirnov
GINI Coefficient is one of methods that have been
used to measure the inequality in population. It
defined as the mean of absolute differences between
all pairs of individuals for some measure.
GINI Coefficient can be applied to measure the
quality of the scorecard. This can be done by
comparing the concentration of “bad” customer on
lower score and “good” customer on higher value.
The objective is to know whether any significant
differrences between the percentage of “good” and
“bad” customer for the same score band.
KS test is one of goodness-of-fit tests. This statistical
test is used to decide if a sample from population
comes from specific distribution. It is useful to
compare between two distributions in population
(Sabato, [10]).
KS test can be also applied to measure scorecard’s
quality. This test is used by comparing the distribution between “good” customer and “bad” customer.
A good scorecard is expected whether the score value
of “bad” customer distribute on lower score rather
than the score value of “good” customer. The differrences between both of distribution indicates that
the quality of the scorecard in discriminate between
“good” and “bad” customer. The difference is reflected
by obtain the KS Score. (https://core.ac.uk/download/pdf/32452521.pdf)